requirements.txt
wheel
pandas
imblearn
numpy
scikit-learn
dvc
xgboost
dill
catboost
lightgbm
mlflow==2.3.*
fastapi
uvicorn
python-dotenv
faker
# -e .